math: Make signgam follow posix and c99 specs
signgam is required by posix and forbidden by c99, so use a hidden
variable '__signgam' to hold the real value and create a weak alias
for 'signgam' which is declared in math.h in posix mode but not in c99
mode.

Finally, we can't make signgam a thread local variable because c99
allows applications to create global variables with that name, and
because posix says it is defined as 'extern int signgam'. Applications
wanting thread safety are required by posix to use lgamma_r.

Signed-off-by: Keith Packard <keithp@keithp.com>