newlib: libc: setenv should consider empty string invalid
According to POSIX, an empty string should be considered as an invalid
name for environment variables, and the implementations should return
-1 and set errno to EINVAL.

https://pubs.opengroup.org/onlinepubs/9699919799/functions/setenv.html

However, both picolibc and newlib support this currently in
setenv. Perhaps it is underspecified in the ISO C standard. Otherwise
it is not really clear to me if this is a feature or a bug.

Recycling the existing pattern from unsetenv(), this should work appropriately

Signed-off-by: Keith Packard <keithp@keithp.com>